feat: add pyright type checking to CI
- Add pyrightconfig.json with basic mode targeting osmose/ and ui/
- Add pyright>=1.1.350 to dev dependencies in pyproject.toml
- Add type-check job to CI workflow (GitHub Actions)
- Fix 36 type errors: genuine fixes (returncode None guard, log10 array
  conversion, normalized_rmse float cast) and targeted suppression of
  framework noise from shiny, pymoo, SALib, sklearn, shiny_deckgl
- All 427 existing tests continue to pass

+++ osmose/calibration/objectives.py
@@ -15,8 +15,8 @@ def _timeseries_rmse(
 ) -> float:
     """Generic RMSE for aligned time series with an optional species filter."""
     if species:
-        simulated = simulated[simulated["species"] == species]
-        observed = observed[observed["species"] == species]
+        simulated = simulated[simulated["species"] == species]  # type: ignore[assignment]
+        observed = observed[observed["species"] == species]  # type: ignore[assignment]
 
     merged = pd.merge(simulated, observed, on="time", suffixes=("_sim", "_obs"))
     if merged.empty:
@@ -91,4 +91,4 @@ def normalized_rmse(simulated: np.ndarray, observed: np.ndarray) -> float:
     if obs_mean == 0:
         return float("inf")
     rmse = float(np.sqrt(np.mean((simulated - observed) ** 2)))
-    return rmse / obs_mean
+    return float(rmse / obs_mean)
